вынес логику обновления в отдельный австрактный класс smart
